Export ClientToken from @y-sweet/react and @y-sweet/client
